+++ packages/react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @feedvalue/react
 
+## 0.1.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @feedvalue/core@0.1.6
+
 ## 0.1.5
 
 ### Patch Changes